+++ lib/plugins/aws/package/compile/events/event-bridge/index.js
