+++ lib/plugins/aws/package/compile/events/event-bridge/utils.js
