+++ lib/plugins/index.js
@@ -49,7 +49,7 @@ module.exports = [
   require('./aws/package/compile/events/cloud-watch-event.js'),
   require('./aws/package/compile/events/cloud-watch-log.js'),
   require('./aws/package/compile/events/cognito-user-pool.js'),
-  require('./aws/package/compile/events/eventBridge/index.js'),
+  require('./aws/package/compile/events/event-bridge/index.js'),
   require('./aws/package/compile/events/sqs.js'),
   require('./aws/package/compile/events/cloud-front.js'),
   require('./aws/package/compile/events/http-api'),
