+++ test/unit/lib/plugins/aws/package/compile/events/event-bridge/index.test.js
